breakout data and components
